Removed extended warnings (#2)
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@17361 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 4318675a0cd5173ac786f42f93b15c162d39fdbe

+++ spatial/rcp.viewer.stream.map/src/de/uniol/inf/is/odysseus/rcp/viewer/stream/map/thematic/buffer/TimeSliderComposite.java
@@ -544,7 +544,7 @@ public class TimeSliderComposite extends Composite implements
 
 		// Set the time-range-label
 		long timeInMs = ((long) rangeSlider.getUpperValue() - (long) rangeSlider
-				.getLowerValue()) * (long) sliderFactor;
+				.getLowerValue()) * sliderFactor;
 		timeRangeLabel.setText(msToTimeString(timeInMs));
 	}
